requirements.txt
# Agent Forge - Python Dependencies

# Core
anthropic>=0.24.0
python-dotenv>=1.0.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Database (optional)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching (optional)
redis>=5.0.0

# Utilities
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4

# Payments
stripe>=7.0.0

# Scheduling
apscheduler>=3.10.0

# ==================== Voice Module ====================

# LiveKit (real-time communication)
livekit>=0.11.0
livekit-agents>=0.8.0

# Telephony
twilio>=9.0.0
# telnyx>=2.0.0  # Alternative telephony provider

# LLM Providers for voice agents
openai>=1.30.0
google-generativeai>=0.5.0

# Speech-to-Text (ASR)
deepgram-sdk>=3.0.0
# assemblyai>=0.20.0  # Alternative ASR provider

# Text-to-Speech (TTS)
elevenlabs>=1.0.0

# WebSocket for real-time audio streaming
websockets>=12.0

# Audio processing
numpy>=1.26.0
scipy>=1.12.0

# Cal.com integration for appointment booking
# (Uses REST API via httpx, no additional package needed)

# ==================== End Voice Module ====================

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Linting
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
